Tidy up files into dirs
