+++ src/skiptracer/plugins/tinder/__init__.py
@@ -3,8 +3,8 @@ from __future__ import absolute_import
 #
 # Tinder Module - illwill
 #
-from plugins.base import PageGrabber
-from .colors import BodyColors as bc
+from ..base import PageGrabber
+#from .colors import BodyColors as bc
 import re
 import logging
 import requests
